♻️ F12: HOOKS > Create useBasket
